feat(core-auth): move XmUserService
