+++ packages/core/index.ts
@@ -7,6 +7,4 @@ export { XmSessionService, ISession } from './src/xm-session.service';
 export { XmCoreModule } from './src/xm-core.module';
 export { XmUiConfigService } from './src/config/xm-ui-config.service';
 export { XmUIConfig } from './src/config/xm-ui-config-model';
-export { XmUser } from './src/auth/xm-user-model';
-export { XmUserService } from './src/auth/xm-user.service';
 export * from './src/xm-core-config';
