+++ packages/core/src/auth/index.ts
@@ -1,2 +1,8 @@
+export { AuthInterceptor } from './src/auth.interceptor';
+export { AuthRefreshTokenService } from './src/auth-refresh-token.service';
+export { XmAuthenticationService} from './src/xm-authentication.service';
+export { XmAuthenticationStoreService } from './src/xm-authentication-store.service';
 export { XmCoreAuthModule } from './src/xm-core-auth.module';
-export * from '../../../../src/app/shared/auth'
+export { XmUserService } from './src/xm-user.service';
+export { XmSessionService } from '../xm-session.service';
+export { XmUser, XmUserLogin, XmUserPermission } from './src/xm-user-model';
