+++ packages/core/src/auth/src/xm-core-auth.module.ts
@@ -1,5 +1,6 @@
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { Injector, ModuleWithProviders, NgModule } from '@angular/core';
+import { XmUserService } from './xm-user.service';
 import { AuthInterceptor } from './auth.interceptor';
 
 @NgModule({
@@ -18,6 +19,7 @@ export class XmCoreAuthModule {
                     multi: true,
                     deps: [Injector],
                 },
+                XmUserService,
             ],
             ngModule: XmCoreAuthModule,
         };
