+++ packages/core/src/xm-core.module.ts
@@ -11,7 +11,6 @@ import { XM_CORE_EXTERNAL_CONFIG, XmCoreConfig } from './xm-core-config';
 import { XmEventManagerService } from './xm-event-manager.service';
 
 import { XmSessionService } from './xm-session.service';
-import { XmUserService } from './auth/xm-user.service';
 
 export function xmCoreConfigFactory(externalConfig?: XmCoreConfig): XmCoreConfig {
     return defaults(externalConfig, new XmCoreConfig());
@@ -29,7 +28,6 @@ export class XmCoreModule {
                 {provide: JhiEventManager, useExisting: XmEventManagerService},
                 XmSessionService,
                 XmUiConfigService,
-                XmUserService,
                 XmPermissionService,
                 {provide: XM_CORE_EXTERNAL_CONFIG, useValue: externalConfig},
                 {provide: XmCoreConfig, useFactory: xmCoreConfigFactory, deps: [XM_CORE_EXTERNAL_CONFIG]},
